refactor(web-ai-api-check): modularize credential extraction
