+++ src/entrypoints/content/webAiApiCheck/components/ApiCheckCandidateButtons.tsx
@@ -1,7 +1,7 @@
 import type { TFunction } from "i18next"
 
 import { cn } from "~/lib/utils"
-import type { ApiCheckCandidate } from "~/services/verification/webAiApiCheck/extractCredentials"
+import type { ApiCheckCandidate } from "~/services/verification/webAiApiCheck/credentialExtraction/candidateContract"
 
 import { WEB_AI_API_CHECK_TEST_IDS } from "../testIds"
