+++ src/features/ApiCredentialProfiles/components/ApiCredentialProfileDialog.tsx
@@ -41,7 +41,7 @@ import { getApiVerificationApiTypeLabel } from "~/services/verification/aiApiVer
 import {
   normalizeGoogleFamilyBaseUrl,
   normalizeOpenAiFamilyBaseUrl,
-} from "~/services/verification/webAiApiCheck/extractCredentials"
+} from "~/services/verification/webAiApiCheck/credentialExtraction/baseUrlCandidates"
 import type { Tag } from "~/types"
 import type {
   ApiCredentialProfile,
